+++ CHANGELOG.md
@@ -35,6 +35,7 @@ release points are not being annotated in GitHub.
 - SIDD file reading in `sarpy/consistency/sidd_consistency.py`
 - Application of adjustable parameter offsets in RIC frames during projection
 - Overflow bug in `ComplexFormatFunction` magnitude/phase -> real/imag
+- Parsing of XBANDS, used when number of image bands > 9
 
 ## [1.3.58] - 2023-08-07
 ### Added